+++ package.json
@@ -54,8 +54,6 @@
     "@hashflow/sdk": "1.2.4",
     "@hashflow/taker-js": "0.0.2",
     "@paraswap/core": "1.1.0",
-    "@sushiswap/router": "^0.0.12",
-    "@sushiswap/viem-config": "^1.0.2",
     "async": "^3.2.4",
     "axios": "0.26.0",
     "bignumber.js": "9.1.0",
