+++ src/dex/uniswap-v3/sushiswap-v3-integration.test.ts
@@ -1,0 +1,111 @@
+/* eslint-disable no-console */
+import dotenv from 'dotenv';
+dotenv.config();
+import { SushiSwapV3 } from '../sushiswap-v3/sushiswap-v3';
+import { Network, SwapSide } from '../../constants';
+import { DummyDexHelper, IDexHelper } from '../../dex-helper';
+import { Tokens } from '../../../tests/constants-e2e';
+import { BI_POWS } from '../../bigint-constants';
+import { checkPoolPrices, checkPoolsLiquidity } from '../../../tests/utils';
+import { Interface, Result } from '@ethersproject/abi';
+import SushiswapV3QuoterV2ABI from '../../abi/sushiswap-v3/QuoterV2.json';
+import { Address } from '@paraswap/core';
+
+const quoterIface = new Interface(SushiswapV3QuoterV2ABI);
+
+function getReaderCalldata(
+  exchangeAddress: string,
+  readerIface: Interface,
+  amounts: bigint[],
+  funcName: string,
+  tokenIn: Address,
+  tokenOut: Address,
+  fee: bigint,
+) {
+  return amounts.map(amount => ({
+    target: exchangeAddress,
+    callData: readerIface.encodeFunctionData(funcName, [
+      [tokenIn, tokenOut, amount.toString(), fee.toString(), 0],
+    ]),
+  }));
+}
+
+function decodeReaderResult(
+  results: Result,
+  readerIface: Interface,
+  funcName: string,
+) {
+  return results.map(result => {
+    const parsed = readerIface.decodeFunctionResult(funcName, result);
+    return BigInt(parsed[0]._hex);
+  });
+}
+
+async function checkOnChainPricing(
+  dexHelper: IDexHelper,
+  sushiSwapV3: SushiSwapV3,
+  funcName: string,
+  blockNumber: number,
+  exchangeAddress: string,
+  prices: bigint[],
+  tokenIn: Address,
+  tokenOut: Address,
+  fee: bigint,
+  _amounts: bigint[],
+) {
+  // Quoter address
+  // const exchangeAddress = '0xb27308f9F90D607463bb33eA1BeBb41C27CE5AB6';
+  const readerIface = quoterIface;
+
+  // const sum = prices.reduce((acc, curr) => (acc += curr), 0n);
+  //
+  // if (sum === 0n) {
+  //   console.log(
+  //     `Prices were not calculated for tokenIn=${tokenIn}, tokenOut=${tokenOut}, fee=${fee.toString()}. Most likely price impact is too big for requested amount`,
+  //   );
+  //   return false;
+  // }
+
+  const readerCallData = getReaderCalldata(
+    exchangeAddress,
+    readerIface,
+    _amounts.slice(1),
+    funcName,
+    tokenIn,
+    tokenOut,
+    fee,
+  );
+
+  let readerResult;
+  try {
+    readerResult = (
+      await dexHelper.multiContract.methods
+        .aggregate(readerCallData)
+        .call({}, blockNumber)
+    ).returnData;
+  } catch (e) {
+    console.log(
+      `Can not fetch on-chain pricing for fee ${fee}. It happens for low liquidity pools`,
+      e,
+    );
+    return false;
+  }
+
+  const expectedPrices = [0n].concat(
+    decodeReaderResult(readerResult, readerIface, funcName),
+  );
+
+  console.log('EXPECTED PRICES: ', expectedPrices);
+
+  let firstZeroIndex = prices.slice(1).indexOf(0n);
+
+  // we skipped first, so add +1 on result
+  firstZeroIndex = firstZeroIndex === -1 ? prices.length : firstZeroIndex;
+
+  // Compare only the ones for which we were able to calculate prices
+  expect(prices.slice(0, firstZeroIndex)).toEqual(
+    expectedPrices.slice(0, firstZeroIndex),
+  );
+  return true;
+}
+
