+++ src/dex/uniswap-v3/uniswap-v3.ts
@@ -91,15 +91,15 @@ export class UniswapV3
 
   public static dexKeysWithNetwork: { key: string; networks: Network[] }[] =
     getDexKeysWithNetwork(
-      _.pick(UniswapV3Config, ['UniswapV3', 'QuickSwapV3.1', 'RamsesV2', 'ChronosV3']),
+      _.pick(UniswapV3Config, ['UniswapV3', 'SushiSwapV3', 'QuickSwapV3.1', 'RamsesV2', 'ChronosV3']),
     );
 
   logger: Logger;
 
-  protected uniswapMulti: Contract;
-  protected stateMultiContract: Contract;
+  private uniswapMulti: Contract;
+  private stateMultiContract: Contract;
 
-  protected notExistingPoolSetKey: string;
+  private notExistingPoolSetKey: string;
 
   constructor(
     protected network: Network,
